backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# GCP
cloud-sql-python-connector[psycopg2]
google-cloud-storage
google-cloud-aiplatform

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Security & Encryption
cryptography==42.0.2

# AI/ML
google-generativeai==0.3.2
